Semi-working Sim (#68)
* Yup

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* please work (#64)

* logs

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Check

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

* Tired

---------

+++ esAPI/Services/BankService.cs
@@ -27,7 +27,7 @@ namespace esAPI.Services
                 var snapshot = new BankBalanceSnapshot
                 {
                     SimulationDay = simulationDay,
-                    Balance = balance,
+                    Balance = (double)balance,
                     Timestamp = simulationDay // Use the day number as timestamp
                 };
                 _db.BankBalanceSnapshots.Add(snapshot);
